requirements.txt
absl-py==1.4.0
accelerate==0.31.0
aiohttp==3.9.3
antlr4-python3-runtime==4.9.3
apex==0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.24.1.1
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
datasets==2.18.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
evaluate==0.3.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
expecttest==0.1.3
fastjsonschema==2.16.3
filelock==3.13.1
fixedpoint==1.0.1
fonttools==4.39.3
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
PuLP==3.1.1
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
grpcio==1.53.0
huggingface-hub==0.21.4
hypothesis==5.35.1
idna==3.4
importlib-metadata==6.3.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.14.5
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
lm-eval==0.3.0
llvmlite==0.39.1
lxml==5.2.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.5
mock==5.0.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.14
murmurhash==1.0.9
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.8.1
notebook==6.4.10
numpy==1.22.2
nvidia-dali-cuda110==1.24.0
nvidia-pyindex==1.0.9
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
peft==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
pluggy==1.0.0
pooch==1.7.0
portalocker==2.8.2
preshed==3.0.8
prettytable==3.7.0
primefac==2.0.12
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==5.26.0
psutil==5.9.8
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycountry==24.6.1
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
pytest-rerunfailures==11.1.2
pytest-shard==0.1.2
pytest-xdist==3.2.1
python-dateutil==2.9.0.post0
python-hostlist==1.23.0
pytorch-quantization==2.1.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.0.2
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.9.3
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sphinx-glpi-theme==0.3
sqlitedict==2.1.0
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminado==0.17.1
thinc==8.1.9
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
transformers==4.36.2
typer==0.7.0
types-dataclasses==0.6.6
typing_extensions==4.5.0
tzdata==2024.1
urllib3==1.26.15
wasabi==1.1.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
xdoctest==1.0.2
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
